to_do_list/requirements.txt
flask==2.2.3
werkzeug==2.2.3
gunicorn==20.1.0
psycopg2-binary==2.9.9
redis==4.5.1
